porting fixes

+++ nossl/md5.cpp
@@ -21,14 +21,14 @@
 #include "md5.h"
 
 #define PUT_64BIT_LE(cp, value) do {                    \
-    (cp)[7] = (value) >> 56;                    \
-    (cp)[6] = (value) >> 48;                    \
-    (cp)[5] = (value) >> 40;                    \
-    (cp)[4] = (value) >> 32;                    \
-    (cp)[3] = (value) >> 24;                    \
-    (cp)[2] = (value) >> 16;                    \
-    (cp)[1] = (value) >> 8;                     \
-    (cp)[0] = (value); } while (0)
+    (cp)[7] = (uint8_t)((value) >> 56);                    \
+    (cp)[6] = (uint8_t)((value) >> 48);                    \
+    (cp)[5] = (uint8_t)((value) >> 40);                    \
+    (cp)[4] = (uint8_t)((value) >> 32);                    \
+    (cp)[3] = (uint8_t)((value) >> 24);                    \
+    (cp)[2] = (uint8_t)((value) >> 16);                    \
+    (cp)[1] = (uint8_t)((value) >> 8);                     \
+    (cp)[0] = (uint8_t)((value)); } while (0)
 
 #define PUT_32BIT_LE(cp, value) do {                    \
     (cp)[3] = (value) >> 24;                    \
@@ -36,7 +36,7 @@
     (cp)[1] = (value) >> 8;                     \
     (cp)[0] = (value); } while (0)
 
-static u_int8_t PADDING[MD5_BLOCK_LENGTH] = {
+static uint8_t PADDING[MD5_BLOCK_LENGTH] = {
     0x80, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
@@ -70,7 +70,7 @@ MD5Update(MD5_CTX *ctx, const unsigned char *input, size_t len)
     need = MD5_BLOCK_LENGTH - have;
 
     /* Update bitcount */
-    ctx->count += (u_int64_t)len << 3;
+    ctx->count += (uint64_t)len << 3;
 
     if (len >= need) {
         if (have != 0) {
@@ -101,7 +101,7 @@ MD5Update(MD5_CTX *ctx, const unsigned char *input, size_t len)
 void
 MD5Pad(MD5_CTX *ctx)
 {
-    u_int8_t count[8];
+    uint8_t count[8];
     size_t padlen;
 
     /* Convert count to 8 bytes in little endian order. */
@@ -151,9 +151,9 @@ MD5Final(unsigned char digest[MD5_DIGEST_LENGTH], MD5_CTX *ctx)
  * the data and converts bytes into longwords for this routine.
  */
 void
-MD5Transform(u_int32_t state[4], const u_int8_t block[MD5_BLOCK_LENGTH])
+MD5Transform(uint32_t state[4], const uint8_t block[MD5_BLOCK_LENGTH])
 {
-    u_int32_t a, b, c, d, in[MD5_BLOCK_LENGTH / 4];
+    uint32_t a, b, c, d, in[MD5_BLOCK_LENGTH / 4];
 
 #if BYTE_ORDER == LITTLE_ENDIAN
     memcpy(in, block, sizeof(in));
